requirements.txt
Flask==2.3.3
Flask-MySQLdb==0.2.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
openai==1.30.1
python-dotenv==1.0.1
